\name{ZIDM-package}
\alias{ZIDM-package}
\alias{ZIDM}
\docType{package}
\title{
\packageTitle{ZIDM}
}
\description{
\packageDescription{ZIDM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ZIDM}
\packageIndices{ZIDM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ZIDM}

Maintainer: \packageMaintainer{ZIDM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
